Bro Parser Improvements
The old Bro parser required that the data came wrapped in a key that
specified the protocol. But not every Bro produces data in that format;
now the parser supports that format AND a format that is flatter and the
protocol is just another attribute in the message.

This also contains some minor code cleanup in the Bro parser.