JSONMapParser should handle unfolding nested maps
The JSONMapParser only supports unfolding of maps at the root of the
JSON object.  It would be better if it could unfold to arbitrary depths
for broader range of support.

This is added by introducing recursion to the UNFOLD functionality.
Because you can’t self reference in the enum lambda, I have added a
private function which can recurse and had the lambda call that.
It is possible that there is a better way to implement this, so
hopefully that will come out in review.