Introduce the ability in the stellar system to differentiate between a variable's presence and its value. That being that there IS a variable at all vs. there is a variable, and its value is NULL.
Referenced variables in Stellar statements with the default variable resolver will throw ParseExceptions
if they are not included or passed.  Other variable resolvers can implement this policy as they need.
For example the LambdaExpression's VariableResolver allows missing variables, since the variable naming
in the lambda expression is arbitrary and separate from the original call context.

The context supports this by maintaining a thread local activity type, such that the context can be inspected
for a validation vs. a parse opteration. Variables are not present during validation and this is allowed.
A thread local variable is used because the context may be used by multiple threads.

Tests have been changed or updated ( hopefully ) to thier _intent_ where possible.  If they were testing
for handling a variable that *is* null vs a missing variable etc.