Adds Variable Assignment to the Stellar Lanaguage
=, +=, -=, *=, /=

This are valid for assignment to variables.  All assignments return the value assigned,
as other scripting languages do.

VariableResolver extended with an update() method, though it may not be supported
by all resolvers, this gives the ability to update variables and save between calls.

The shell was updated for this, such that assignment calls

More testing need and doc but I wanted to get it out there.
This is dependent on PR 686