Remove description of Global configuration
It confuses reader. Description of Global configuration is available ahead of the enrichment and listed here:
https://metron.apache.org/current-book/metron-platform/metron-common/index.html
Global config is used for validation of the fields.
Enrichment is used for fields enrichment and probably can also be used for validation of the fields (which is not covered in the doc).
From available doc I can see that they serve two different purposes.

So, I don't see any reason to mention global config in enrichment description. What's the benefit of knowing about global configuration when you want deep dive in enrichment?