Stellar shell functionality to verify stellar statements.
This will allow users to check their deployed statements, say after upgrade, when they are at rest ( and would fail on use ).
In other words, they were valid when stored, but are not now because of stellar changes, such as new keywords.

The interface StellarConfiguredStatementReporter, which is @IndexSubclasses marked, allows the shell to discover
reporters that can provide statements for validation.  This discovery allows de-coupling of stellar and 'hosts' that
know about the location of the stored statements, and the configuration structure details.

We do mention the configurations in the shell output at this time.

metron-common implements this interface, and can run through visiting all the configurations.